services/worker-python/requirements.txt
psycopg[binary]>=3.1.18
python-dotenv>=1.0.1
requests>=2.32.0
qdrant-client>=1.10.1
